requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
twilio==9.2.3
openai>=1.23.0
